int to int64

+++ kpp.go
@@ -70,7 +70,7 @@ func getHref(body []byte, reStr string) (string, error) {
 }
 
 // GetRating - получение рейтингов
-func GetRating(name string, engName string, year int) (KP, error) {
+func GetRating(name string, engName string, year int64) (KP, error) {
 	var (
 		kp   KP
 		href string
